move SystemUserConfigurer to configuration package
